get things working for web/vm tests
