+++ pkgs/test_core/lib/src/runner/runner_test.dart
@@ -71,7 +71,7 @@ class RunnerTest extends Test {
             // When we kill the isolate that the test lives in, that will close
             // this virtual channel and cause the spawned isolate to close as
             // well.
-            spawnHybridUri(message['url'] as String, message['message'])
+            spawnHybridUri(message['url'] as String, message['message'], suite)
                 .pipe(testChannel.virtualChannel(message['channel'] as int));
             break;
         }
